Updated version for a new minor and AppHeader relocated
