+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "desde-los-margenes-movieguide",
-  "version": "0.0.3",
+  "version": "0.1.0",
   "author": "Luis Gil Gutiérrez de la Barreda",
   "private": true,
   "scripts": {
