+++ server/package.json
@@ -1,6 +1,6 @@
 {
   "name": "server",
-  "version": "0.0.3",
+  "version": "0.1.0",
   "description": "Server for desde-los-margenes-movieguide",
   "main": "server.js",
   "dependencies": {
